% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTargetDecoys.R
\name{evalTargetDecoys}
\alias{evalTargetDecoys}
\title{Evaluate assumptions of the Target Decoys approach}
\usage{
evalTargetDecoys(object, decoy = NULL, score = NULL, log10 = NULL, nBins = 50)
}
\arguments{
\item{object}{A \code{data.frame}, \linkS4class{mzID} or \linkS4class{mzRident} object.}

\item{decoy}{\code{character}, name of the variable that indicates if the peptide
matches to a target or to a decoy.}

\item{score}{\code{numeric}, indicating the score of the peptide match, obtained
by the search engine.}

\item{log10}{\code{logical} to indicate if the score should be
\code{-log10}-transformed.}

\item{nBins}{\code{numeric} indicating the number of bins in the histogram.}
}
\value{
A list containing the PP-plot and histogram, a zoom of both plots and an
overview of all four plots.
}
\description{
Create diagnostic plots to evaluate the TDA assumptions.
A histogram and PP plot allow to check both necessary assumptions.
}
\examples{
library(mzID)

## Use one of the example files in the mzID package
exampleFile <- system.file('extdata', '55merge_tandem.mzid', package = 'mzID')
mzIDexample <- mzID(exampleFile)

decoyPlots <- evalTargetDecoys(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE
)

# Plot only the overview of the four plots
decoyPlots$together

# If you do not know the name of the score and/or decoy variable,
# or if you want to evaluate how many bins you want to use in the histogram,
# or if -log10 transformation is needed you can launch a shiny app by only
# specifying the mzID object

# evalTargetDecoys(mzIDexample)


## mzRident objects can also be used
library(mzR)

if (requireNamespace("msdata", quietly = TRUE)) {
   ## Using example file from msdata
   file <- system.file("mzid", "Tandem.mzid.gz", package="msdata")
   mzid <- openIDfile(file)
}
decoyPlots2 <- evalTargetDecoys(mzid,
    decoy = "isDecoy", score = "X.Tandem.expect", log10 = TRUE
)
decoyPlots2$together
}
\author{
Elke Debrie, Lieven Clement
}
