% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModSwissXT}
\alias{ModSwissXT}
\title{Swiss-Prot X!Tandem data}
\format{
An \linkS4class{mzID} object.
}
\usage{
data(ModSwissXT)
}
\description{
Data from a Pyrococcus furiosis sample run on a LTQ-Orbitrap Velos mass spectrometer.
The data can be found in the PRIDE repository with identifier PXD001077.
The Pyrococcus furiosis reference proteome fasta files were downloaded from
UniProtKB/Swiss-Prot on April 22, 2016.
The Pyrococcus data was searched against all Pyrococcus proteins
with a combined search (omssa, X!Tandem and MS-GF+) using the reference proteome
from UniProtKB/Swiss-Prot.
}
\keyword{datasets}
