% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyroSwissprot}
\alias{pyroSwissprot}
\title{SwissProt data with different search engines}
\format{
An mzID object.
TODO: add some more details...
}
\source{
TODO: add source of dataset
}
\usage{
data(pyroSwissprot)
}
\description{
TODO: describe data briefly...
}
\keyword{datasets}
