% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TargetDecoy-package.R
\docType{package}
\name{TargetDecoy-package}
\alias{TargetDecoy}
\alias{TargetDecoy-package}
\title{TargetDecoy: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/statOmics/TargetDecoy}
  \item Report bugs at \url{https://github.com/statOmics/TargetDecoy/issues}
}

}
\author{
\strong{Maintainer}: Milan Malfait \email{milan.malfait94@gmail.com} (\href{https://orcid.org/0000-0001-9144-3701}{ORCID})

}
\keyword{internal}
