% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTargetDecoys.R
\name{decoyScoreTable}
\alias{decoyScoreTable}
\title{Prepare score table for decoys}
\usage{
decoyScoreTable(object, decoy = NULL, score = NULL, log10 = TRUE)
}
\arguments{
\item{object}{A \code{data.frame}, \linkS4class{mzID} or \linkS4class{mzRident} object.}

\item{decoy}{\code{character}, name of the variable that indicates if the peptide
matches to a target or to a decoy.}

\item{score}{\code{numeric}, indicating the score of the peptide match, obtained
by the search engine.}

\item{log10}{\code{logical} to indicate if the score should be
\code{-log10}-transformed.}

\item{nBins}{\code{numeric} indicating the number of bins in the histogram.}
}
\value{
A \code{data.frame} with a logical \code{"decoy"} column and numeric \code{"scores"}.
}
\description{
Takes an input object and returns a score table for the decoys. If one of the
arguments besides \code{object} is missing, will open a Shiny app to interactively
select the variables.
}
\author{
Elke Debrie, Lieven Clement
}
\keyword{internal}
