% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalTargetDecoys.R
\name{evalTargetDecoys}
\alias{evalTargetDecoys}
\alias{evalTargetDecoysPPPlot}
\alias{evalTargetDecoysHist}
\title{Evaluate assumptions of the Target Decoys approach}
\usage{
evalTargetDecoys(object, decoy, score, log10 = TRUE, nBins = 50)

evalTargetDecoysPPPlot(
  object,
  decoy,
  score,
  log10 = TRUE,
  nBins = 50,
  zoom = FALSE
)

evalTargetDecoysHist(
  object,
  decoy = NULL,
  score = NULL,
  log10 = NULL,
  nBins = 50,
  zoom = FALSE
)
}
\arguments{
\item{object}{A \code{data.frame}, \linkS4class{mzID} or \linkS4class{mzRident} object.}

\item{decoy}{\code{character}, name of the variable that indicates if the peptide
matches to a target or to a decoy.}

\item{score}{\code{numeric}, indicating the score of the peptide match, obtained
by the search engine.}

\item{log10}{\code{logical} to indicate if the score should be
\code{-log10}-transformed.}

\item{nBins}{\code{numeric} indicating the number of bins in the histogram.}

\item{zoom}{Logical value indicating whether a zoomed version of the plot
should be returned. Default: \code{FALSE}.}
}
\value{
\code{evalTargetDecoys} returns an overview of the following four plots:
\enumerate{
\item A PP plot showing the empirical cumulative distribution of the target
distribution in function of that of the decoy distribution
\item A histogram showing the score distributions of the decoys and non-decoys
\item A zoomed PP plot
\item A zoomed histogram
}

\code{evalTargetDecoysPPPlot} generates the PP plot only (1.) or the zoomed
version (3.) if \code{zoom = TRUE}.

\code{evalTargetDecoysHist} generates the histogram only (2.) or the zoomed
version (4.) if \code{zoom = TRUE}.
}
\description{
Create diagnostic plots to evaluate the TDA assumptions.
A histogram and PP plot allow to check both necessary assumptions.
}
\examples{
library(mzID)

## Use one of the example files in the mzID package
exampleFile <- system.file("extdata", "55merge_tandem.mzid", package = "mzID")
mzIDexample <- mzID(exampleFile)

# Plot the overview of the four plots
evalTargetDecoys(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE
)

# Plot the PP plot only
evalTargetDecoysPPPlot(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE
)

# Plot the zoomed PP plot only
evalTargetDecoysPPPlot(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE,
    zoom = TRUE
)

# Plot the histogram only
evalTargetDecoysHist(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE
)

# Plot the zoomed histogram only
evalTargetDecoysHist(mzIDexample,
    decoy = "isdecoy", score = "x\\\\!tandem:expect", log10 = TRUE,
    zoom = TRUE
)

## mzRident objects can also be used
library(mzR)

if (requireNamespace("msdata", quietly = TRUE)) {
    ## Using example file from msdata
    file <- system.file("mzid", "Tandem.mzid.gz", package = "msdata")
    mzid <- openIDfile(file)
}
evalTargetDecoys(mzid,
    decoy = "isDecoy", score = "X.Tandem.expect", log10 = TRUE
)
}
\author{
Elke Debrie, Lieven Clement, Milan Malfait
}
