% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModSwissXT}
\alias{ModSwissXT}
\title{SwissProt X!Tandem data}
\format{
An mzID object.
TODO: add some more details...
}
\source{
TODO: add source of dataset
}
\usage{
data(ModSwissXT)
}
\description{
TODO: describe data briefly...
}
\keyword{datasets}
